modularize and improves
